pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "freewill"
version = "0.1.0"
description = "Autonomous LLM System"
requires-python = ">=3.9,<3.10"
dependencies = [
    "Flask==3.0.0",
    "pytest",
    "llama-cpp-python==0.2.11",
    "pyttsx3==2.90",
    "aiosqlite==0.19.0",
    "psutil==5.9.6",
    "flask-socketio==5.3.6",
    "python-socketio==5.8.0",
    "werkzeug==3.0.1",
    "typing-extensions>=4.8.0",  # Updated to be compatible with torch
    "python-dotenv==1.0.0",
    "torch>=2.5.1"  # Explicitly include torch
]

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.pytest.ini_options]
pythonpath = [
    "src",
    "tests"
]
testpaths = ["tests"]

[tool.pylance]
reportMissingImports = "warning"
python.analysis.extraPaths = ["src"]